feat: implement chunk_gated_delta_rule fused operator for FlagGems
PyTorch-native + Triton hybrid implementation of chunk_gated_delta_rule
for Qwen3-Next model, with cross-platform support.

Triton path (NVIDIA, K<=128): Uses optimized fused kernels with chunk-based
processing, WY representation, and autotuning for maximum performance.

Native path (Iluvatar or K>128): Falls back to PyTorch-native implementation
for compatibility with hardware that has Triton limitations.

Key features:
- Chunk-based processing with WY representation for linear complexity
- Full dtype support: float32, float16, bfloat16
- Variable-length sequences via cu_seqlens
- GQA support (Hg < H)
- 68 test cases covering all edge cases
- Hybrid dispatch for optimal performance across platforms

Files:
- src/flag_gems/fused/FLA/chunk_gated_delta_rule_native.py (native implementation)
- src/flag_gems/fused/FLA/chunk.py (hybrid dispatch logic)
- src/flag_gems/fused/FLA/wy_fast.py (fix fp16/bf16 exp)
- src/flag_gems/fused/FLA/chunk_delta_h.py (fix fp16/bf16 exp)
- src/flag_gems/fused/FLA/chunk_o.py (fix fp16/bf16 exp)
- src/flag_gems/fused/FLA/fused_cumsum_kkt_solve_tril.py (fix fp16/bf16 exp)
- tests/test_FLA/test_chunk_gated_delta_rule.py (68 tests)
- benchmark/test_FLA/test_chunk_gated_delta_rule_perf.py

+++ src/flag_gems/fused/FLA/chunk_delta_h.py
@@ -188,8 +188,8 @@ def chunk_gated_delta_rule_fwd_kernel_h_blockdim64(
                 g + bos * H + i_h, (T,), (H,), (i_t * BT,), (BT,), (0,)
             )
             b_g = tl.load(p_g, boundary_check=(0,))
-            b_v = b_v * tl.where(m_t, exp(b_g_last - b_g), 0)[:, None]
-            b_g_last = exp(b_g_last)
+            b_v = b_v * tl.where(m_t, exp((b_g_last - b_g).to(tl.float32)), 0)[:, None]
+            b_g_last = exp(b_g_last.to(tl.float32))
             b_h1 *= b_g_last
             if K > 64:
                 b_h2 *= b_g_last
@@ -205,7 +205,7 @@ def chunk_gated_delta_rule_fwd_kernel_h_blockdim64(
                 mask=(o_k1 < K),
                 other=0.0,
             )
-            b_h1 *= exp(b_gk_last1)[:, None]
+            b_h1 *= exp(b_gk_last1.to(tl.float32))[:, None]
             if K > 64:
                 o_k2 = 64 + o_k1
                 b_gk_last2 = tl.load(
@@ -213,7 +213,7 @@ def chunk_gated_delta_rule_fwd_kernel_h_blockdim64(
                     mask=(o_k2 < K),
                     other=0.0,
                 )
-                b_h2 *= exp(b_gk_last2)[:, None]
+                b_h2 *= exp(b_gk_last2.to(tl.float32))[:, None]
             if K > 128:
                 o_k3 = 128 + o_k1
                 b_gk_last3 = tl.load(
@@ -221,7 +221,7 @@ def chunk_gated_delta_rule_fwd_kernel_h_blockdim64(
                     mask=(o_k3 < K),
                     other=0.0,
                 )
-                b_h3 *= exp(b_gk_last3)[:, None]
+                b_h3 *= exp(b_gk_last3.to(tl.float32))[:, None]
             if K > 192:
                 o_k4 = 192 + o_k1
                 b_gk_last4 = tl.load(
@@ -229,7 +229,7 @@ def chunk_gated_delta_rule_fwd_kernel_h_blockdim64(
                     mask=(o_k4 < K),
                     other=0.0,
                 )
-                b_h4 *= exp(b_gk_last4)[:, None]
+                b_h4 *= exp(b_gk_last4.to(tl.float32))[:, None]
         b_v = b_v.to(k.dtype.element_ty)
 
         p_k = tl.make_block_ptr(
